#101 Move LiveTracker components into grouped dirs.
