+++ app/debugger/modules/variableHijacker/components/subcomponents/NumberSliderRange.tsx
