+++ app/debugger/modules/variableHijacker/components/subcomponents/NumericInput.tsx
