+++ app/debugger/modules/variableHijacker/components/subcomponents/NumericSlider.tsx
