+++ app/debugger/modules/variableHijacker/components/subcomponents/TextInput.tsx
