+++ app/debugger/modules/variableHijacker/components/variableControl/gizmoMap.ts
